ml_models/requirements.txt
tensorflow>=2.12.0
pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
sounddevice>=0.4.6
openai-whisper>=20230918
torch>=2.0.0